Load calibration from remote hash-mapped json file
Json calibration files can now optionally be kept in MapTiles/Mercator/Calibration, decoupled from charts.
This directory is scanned at startup by ChartService and a std::Map created of sha256 hash -> json file.
During load of calibration for a chart in PDFSource, if a name-matched co-located json or kml file is not found, then a hash-matched json file is used.