Introduce a container with a custom clickable list
Introduce a new ContainerWithClickableCustomList class that is
instantiated with a custom window title, a list of string items
(via the setListItems method), a cancel, and a select callback that
is called every time the user clicks on a list item. The select
callback takes as input the index of the clicked item in the list
and it will perform a user-defined action based on the index. It's
up to the user to perform the correct action for a given index.

This commit also binds the new ContainerWithClickableCustomList to
the "Online" button that was introduced in the previous commit.
Currently, the container has two fixed entries namely "Dummy map 1"
and "Dummy map 2", and whenever the user clicks any of the entries,
a message with the corresponding name that was clicked in the list
is logged in the console. This is just to demonstrate how the new
container works.

In the next commit I will introduce the new slippy map config for
the online maps.