Add support for optional user defined comms protocol in mapconfig.json
Add a new `protocol` configuration option in the map definitions in
mapconfig.json. This defines the communication protocol that is spoken
by the tile servers. The `protocol` config option is optional, and if
skipped, `https` is assumed by default. The only accepted values are
`https` and `http`.

Most of the public servers nowadays speak HTTPS, but still some use
HTTP (without the S). Moreover, when using a local test tile server,
setting up a valid SSL certificate can be a pain, so adding support for
HTTP servers compliments the "custom online slippy tiles" feature
nicely.