Visual menu feedback about the enabled online map
The "Online" maps button brings up a menu and allows you to select from
one or more maps. With this commit we store the name of the selected
online map when successfully applied, and show the name of the selected
map in the onlineMapsLabel next time the user clicks on the button that
shows the map settings menu.