Optimisation of map overlay drawing and related modificatons
Filter reported nodes within the visit function so that only configured nodes are processed.
Keep overlay nodes after frame has been drawn, and reuse them in the next frame if they are still required.
Make decisions on what to draw, and what detail to draw, based on the map density rather than the width.
Do not select node for highlighting when map is being panned, only when clicked (a small amount of movement is considered to be a click).