Fix Plane Manual app file browsing initialisation.
This change fixes the bug in the Plane app, which attempts to start
the file browser in a subdirectory of the aircraft directory when
looking for manuals for the current aircraft.

The browse start directory was being initialised too early, before the
aircraft was loaded. This is now done whenever the plane is loaded, so
will work correctly if a new plane is selected during a simulation run.