Convert raw pointers to LoadManager into shared pointers.
Following discussions in a separate PR, change the xdata loaders to
be provided with, and to store, a shared pointer to the LoadManager,
rather than the raw pointer that was previously given. This is mainly
a C++ style concern, since the loaders have a relatively short lifespan
and never outlive the LoadManager.