Refactoring of Route and RouteFinder classes.
Removal of embedded struct RouteDirection from RouteFinder and insertion into Route as Route::Leg.
This is the more logical name and place for the struct as it is a component part of a Route and has no specific relationship with RouteFinder.
Addition of a world API function to provide a RouteFinder instance. (This will make more sense when the SQL alternative to XWorld is added.)