Update curl submodule to v8.7.1.
curl-8.7.1 is the most recent version that does not require an update
to mbedtls.

Added bcrypt to the list of libraries required by the MinGW build
system. Other platforms appear to handle this automatically.

Downgraded the required version of CMake to 3.16 since this is the version
currently supported by Ubuntu/WSL and the Avitab CMake scripts do not
require any features that are not available in 3.16.