Fix issues with query strings in mapconfig.json
On windows, query strings in tile server URLs can cause crashes. So catch exceptions in Platform.cpp to protect against bad filenames
Tile server URLs with query strings can't be used as a tile disk cache filename. So massage the tileURL to be a legal filename. And change dir separators to underscores to reduce the amount of dir creation disk operations.